+++ src/MIDI.jl
@@ -30,7 +30,7 @@ export BPM, bpm, qpm, time_signature, tempochanges, ms_per_tick
 export getnotes, addnote!, addnotes!, addevent!, addevents!
 export MIDITrack
 export Note, Notes, AbstractNote, DrumNote
-export pitch_to_name, name_to_pitch
+export pitch_to_name, name_to_pitch, is_octave
 export pitch_to_hz, hz_to_pitch
 export TrackEvent, MetaEvent, MIDIEvent, SysexEvent
 export readvariablelength, writevariablelength