+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+# v2.3.0
+* New functions `is_octave`.
+* Implement `Base.keys(::Notes)` and `Base.eachindex(::Notes)`.
 # v2.2.0
 * New functions `pitch_to_hz ` and `hz_to_pitch `
 # v2.1.0